requirements.txt
pandas
ta
python-binance
flask
